# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: ERP Barrio API
  version: 0.3.8
  description: Contrato inicial del API del MVP ERP.
servers:
  - url: http://localhost:8000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
...
    EmployeeCreateRequest:
      type: object
      required: [employee_code, full_name, role]
      properties:
        employee_code:
          type: string
          minLength: 1
        full_name:
          type: string
          minLength: 1
        role:
          type: string
          minLength: 1
        is_active:
          type: boolean
          default: true
    EmployeeUpdateRequest:
      type: object
      properties:
        full_name:
          type: string
          minLength: 1
        role:
          type: string
          minLength: 1
        is_active:
          type: boolean
    EmployeeResponse:
      type: object
      required: [id, employee_code, full_name, role, is_active]
      properties:
        id:
          type: string
        employee_code:
          type: string
        full_name:
          type: string
        role:
          type: string
        is_active:
          type: boolean
    EmployeeListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeResponse'
...
    PaymentCreateRequest:
      type: object
      required: [sale_id, amount, method, status, idempotency_key]
      properties:
        sale_id:
          type: string
          minLength: 1
        amount:
          type: number
          exclusiveMinimum: 0
        method:
          type: string
          minLength: 1
        status:
          type: string
          minLength: 1
        idempotency_key:
          type: string
          minLength: 1
    PaymentUpdateRequest:
      type: object
      properties:
        status:
          type: string
          minLength: 1
    PaymentResponse:
      type: object
      required: [id, sale_id, amount, method, status, idempotency_key]
      properties:
        id:
          type: string
        sale_id:
          type: string
        amount:
          type: number
        method:
          type: string
        status:
          type: string
        idempotency_key:
          type: string
    PaymentListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PaymentResponse'
...
paths:
...
  /employees:
    get:
      tags: [employees]
      summary: Lista colaboradores del ERP
      operationId: getEmployees
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Listado de colaboradores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeListResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [employees]
      summary: Crea un colaborador del ERP
      operationId: postEmployees
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreateRequest'
      responses:
        '201':
          description: Colaborador creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Código de colaborador duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /employees/{employee_id}:
    parameters:
      - in: path
        name: employee_id
        required: true
        schema:
          type: string
    get:
      tags: [employees]
      summary: Obtiene un colaborador por identificador
      operationId: getEmployeeById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Colaborador encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Colaborador no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [employees]
      summary: Actualiza campos de un colaborador
      operationId: patchEmployeeById
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdateRequest'
      responses:
        '200':
          description: Colaborador actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Colaborador no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [employees]
      summary: Elimina un colaborador del ERP
      operationId: deleteEmployeeById
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Colaborador eliminado
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Colaborador no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /payments:
    get:
      tags: [payments]
      summary: Lista pagos registrados
      operationId: getPayments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Listado de pagos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentListResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [payments]
      summary: Crea un pago
      operationId: postPayments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreateRequest'
      responses:
        '201':
          description: Pago creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: idempotency_key duplicada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /payments/{payment_id}:
    parameters:
      - in: path
        name: payment_id
        required: true
        schema:
          type: string
    get:
      tags: [payments]
      summary: Obtiene un pago por identificador
      operationId: getPaymentById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pago encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Pago no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [payments]
      summary: Actualiza estado de un pago
      operationId: patchPaymentById
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentUpdateRequest'
      responses:
        '200':
          description: Pago actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Pago no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [payments]
      summary: Elimina un pago
      operationId: deletePaymentById
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Pago eliminado
        '401':
          description: Token ausente/inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Rol insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Pago no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...